Module aws_sdk_iotthingsgraph::operation

Structs

AssociateEntityToThing

Associates a device with a concrete thing that is in the user's registry.

CreateFlowTemplate

Creates a workflow template. Workflows can be created only in the user's namespace. (The public namespace contains only

CreateSystemInstance

Creates a system instance.

CreateSystemTemplate

Creates a system. The system is validated against the entities in the

DeleteFlowTemplate

Deletes a workflow. Any new system or deployment that contains this workflow will fail to update or deploy.

DeleteNamespace

Deletes the specified namespace. This action deletes all of the entities in the namespace. Delete the systems and flows that use entities in the namespace before performing this action.

DeleteSystemInstance

Deletes a system instance.

DeleteSystemTemplate

Deletes a system. New deployments can't contain the system after its deletion.

DeploySystemInstance

DeprecateFlowTemplate

Deprecates the specified workflow. This action marks the workflow for deletion. Deprecated flows can't be deployed, but existing deployments will continue to run.

DeprecateSystemTemplate

Deprecates the specified system.

DescribeNamespace

Gets the latest version of the user's namespace and the public version that it is tracking.

DissociateEntityFromThing

Dissociates a device entity from a concrete thing. The action takes only the type of the entity that you need to dissociate because only

GetEntities

Gets definitions of the specified entities. Uses the latest version of the user's namespace by default. This API returns the

GetFlowTemplate

Gets the latest version of the DefinitionDocument and FlowTemplateSummary for the specified workflow.

GetFlowTemplateRevisions

Gets revisions of the specified workflow. Only the last 100 revisions are stored. If the workflow has been deprecated,

GetNamespaceDeletionStatus

Gets the status of a namespace deletion task.

GetSystemInstance

Gets a system instance.

GetSystemTemplate

Gets a system.

GetSystemTemplateRevisions

Gets revisions made to the specified system template. Only the previous 100 revisions are stored. If the system has been deprecated, this action will return

GetUploadStatus

Gets the status of the specified upload.

ListFlowExecutionMessages

Returns a list of objects that contain information about events in a flow execution.

ListTagsForResource

Lists all tags on an AWS IoT Things Graph resource.

SearchEntities

Searches for entities of the specified type. You can search for entities in your namespace and the public namespace that you're tracking.

SearchFlowExecutions

Searches for AWS IoT Things Graph workflow execution instances.

SearchFlowTemplates

Searches for summary information about workflows.

SearchSystemInstances

Searches for system instances in the user's account.

SearchSystemTemplates

Searches for summary information about systems in the user's account. You can filter by the ID of a workflow to return only systems that use the specified workflow.

SearchThings

Searches for things associated with the specified entity. You can search by both device and device model.

TagResource

Creates a tag for the specified resource.

UndeploySystemInstance

Removes a system instance from its target (Cloud or Greengrass).

UntagResource

Removes a tag from the specified resource.

UpdateFlowTemplate

Updates the specified workflow. All deployed systems and system instances that use the workflow will see the changes in the flow when it is redeployed. If you don't want this

UpdateSystemTemplate

Updates the specified system. You don't need to run this action after updating a workflow. Any deployment that uses the system will see the changes in the system when it is redeployed.

UploadEntityDefinitions

Asynchronously uploads one or more entity definitions to the user's namespace. The document parameter is required if